requirements.txt
pylint==2.9.3
pytest
click
black==20.8b1
pytest-cov
#tabula-py==2.3.0
tabula-py==0.9.0
camelot-py[cv]
PyPDF2
pyspark